requirements.txt
altair==5.0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.1.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.2
cryptography==41.0.3
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.4
django-allauth==0.56.1
django-cors-headers==4.2.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
entrypoints==0.4
executing==1.2.0
extra-streamlit-components==0.1.60
fastjsonschema==2.16.2
filelock==3.12.4
firebase-admin==6.3.0
fqdn==1.5.1
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-status==1.60.0
gspread==5.12.0
httplib2==0.22.0
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
iniconfig==2.0.0
inotify==0.2.10
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jwcrypto==1.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
msgpack==1.0.7
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.9.8
platformdirs==2.6.0
pluggy==1.3.0
poetry-core==1.7.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.24.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
PyMuPDF==1.23.19
PyMuPDFb==1.23.9
pynvim==0.4.3
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==3.0.1
pyproject_hooks==1.0.0
Pyrebase4==4.7.1
pyrsistent==0.19.2
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
python-jwt==4.1.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywifi==1.1.12
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rapidfuzz==3.5.1
reportlab==4.0.8
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rsa==4.9
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.25.0
streamlit-authenticator==0.2.3
streamlit-toggle-switch==1.0.2
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.2
traitlets==5.8.0
trove-classifiers==2023.10.18
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.18
validators==0.21.2
virtualenv==20.24.5
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
wrapt==1.16.0
zipp==3.16.2
